apres test

+++ STG/PROJET/bae/generate_random_instance.py
@@ -25,9 +25,9 @@ def generate_random_instance(file_path: str, nb_periods: int, nb_items: int, mar
         Dt = np.cumsum(dt)
 
         # Costs
-        cB: float = random.uniform(0, 10) + margin_settings
-        cI: float = random.uniform(0, 10) + margin_settings
-        bP: float = random.uniform(0, 10) + margin_settings
+        cB: int = int(random.uniform(0, 10) + margin_settings)
+        cI: int = int(random.uniform(0, 10) + margin_settings)
+        bP: int = int(random.uniform(0, 10) + margin_settings)
         f.write(f"{cB} {cI} {bP}\n")
 
         X = np.zeros(nb_periods, dtype=int)
